% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmed.R
\name{DMR_built}
\alias{DMR_built}
\title{Build DMR vector}
\usage{
DMR_built(res, methylation, nb_cpg = 2)
}
\arguments{
\item{res}{result object of DMR_search function}

\item{methylation}{chromosomal position of markers (start)}

\item{nb_cpg}{chromosomal position of markers (end)}
}
\value{
A set of build DMRs.

DMR_acp contains the first components of each PCA for each DMR.
CpG_for_each_DMR contains the list of markers (CpGs) present on each DMR.
}
\description{
To build a vector for each DMR find with DMR_search
}
\details{
We use the series of pValues (one pValue per CpGs) obtained with the EWASmultivariate
regression method and the combination of pValue max2.
To determine the potential DMRs used the combp method present in the ENmix package (Xu et al. 2016).
This method uses the Fisher method to combine the pValues and also the base pair distance (bP)
between CpGs (1000 bP maximum between nb_cpg CpGs on the same DMR).
The information for each DMR is summarized by running a PCA by DMR on all of the CpGs present on each DMR.
Recovering the first principal component of each PCA,
we therefore have a vector corresponding to the first principal component of a PCA for each DMR.
}
\examples{

}
\author{
Basile Jumentier
}
