% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmed.R
\name{DMR_search}
\alias{DMR_search}
\title{Find DMR}
\usage{
DMR_search(chr, start, end, pval, cpg, ...)
}
\arguments{
\item{chr}{chromosomes}

\item{start}{chromosomal position of markers (start)}

\item{end}{chromosomal position of markers (end)}

\item{pval}{pValues for each markers, from the max2 function}

\item{cpg}{name of each markers}

\item{...}{argument of the combp function from the ENmix package}
}
\value{
A set of selected DMRs.
}
\description{
To identify differentially methylated regions using a modified comb-p method.
}
\details{
The function will use a modified comb-p method to identify
differentially methylated regions (DMRs).
}
\examples{

library(highmed)

# Run multivariate EWAS

res <- multivariate_EWAS(X = example$X, Y = example$Y, M = example$M, K = 5)

# Keep latent factors for univariate mediation

U <- res$U

# Run max2

res <- max2(pval1 = res$pValue[, 1], pval2 = res$pValue[, 2])

# lauch DMR_search

res <- DMR_search(chr = example$annotation$chr,
                  start = example$annotation$start,
                  end = example$annotation$end,
                  pval = res$pval,
                  cpg = example$annotation$cpg, nCores = 1)

}
\references{

}
\author{
Basile Jumentier
}
