% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{Dataset for example}
\format{A list with 4 objects
\describe{
  \item{M}{methylation matrix, 668 individuals and 10000 probes}
  \item{X}{Exposure for 668 individuals}
  \item{Y}{Ouctome for 668 individuals}
  \item{annotation}{Annotation for the 10000 probes}
}}
\usage{
example
}
\description{
Dataset for example
}
\keyword{datasets}
