% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmed.R
\name{max2}
\alias{max2}
\title{max2 Compute the squared maximum of two series of pValues}
\usage{
max2(pval1, pval2, diagnostic.plot = F, ...)
}
\arguments{
\item{pval1}{vector of pValues (p*1) of exposure.}

\item{pval2}{vector of pValues (p*1) of ouctome.}

\item{diagnostic.plot}{if TRUE the histogram of the p-values together
with the estimate of eta0 null line is plotted.
This is useful to visually check the fit of the estimated proportion of null p-values.}

\item{...}{argument of the fdrtool function from the fdrtool package}
}
\value{
an object with the following attributes:

 - a pValue for each markers

 - a qValue for each markers

 - the eta0 of the set of pValues
}
\description{
This function compute the squared maximum of two series of pValues from the multivariate_EWAS() function.
The objective of this function is to test all the markers and to determine which could be
potential mediators in the exposure-outcome association.
}
\details{
The pValue is computed for each markers following this formula

\deqn{pV = max(pVal1, pVal2)^2}

This quantity eta0, i.e. the proportion eta0 of null p-values in a given vector of p-values,
is an important parameter when controlling the false discovery rate (FDR).
A conservative choice is eta0 = 1 but a choice closer to the true value will
increase efficiency and power - see Benjamini and Hochberg (1995, 2000) and Storey (2002) for details.
We use the fdrtool package to transform pValues into qValues,
which allows us to control the FDR.
}
\examples{

}
\author{
Basile Jumentier
}
