% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmed.R
\name{multivariate_EWAS}
\alias{multivariate_EWAS}
\title{multivariate_EWAS}
\usage{
multivariate_EWAS(X, Y, M, K, covar = NULL)
}
\arguments{
\item{X}{an explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Exposure).
Explanatory variables must be encoded as numeric variables.}

\item{Y}{an explanatory variable matrix with n rows and d columns.
Each column corresponds to a distinct explanatory variable (Outcome).
Explanatory variables must be encoded as numeric variables.}

\item{M}{a response variable matrix with n rows and p columns.
Each column corresponds to a distinct response variable (beta-normalized methylation profile).
Response variables must be encoded as numeric.}

\item{K}{an integer for the number of latent factors in the regression model.}
}
\value{
an object of class \code{lfmm} with the following attributes:
 - U the latent variable score matrix with dimensions n x K,
 - V the latent variable axis matrix with dimensions p x K,
 - B the effect size matrix with dimensions p x d.
}
\description{
This function computes regularized least squares estimates
for latent factor mixed models using a ridge penalty.
}
\details{
The response variable matrix Y and the explanatory variable are centered.
}
\examples{

}
\references{

}
\author{
Basile Jumentier
}
