% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmed.R
\name{plot_summary_ACME}
\alias{plot_summary_ACME}
\title{Summary plot for ACME}
\usage{
plot_summary_ACME(ACME)
}
\arguments{
\item{ACME}{the table of ACME from the univariate_mediation() function}
}
\value{
Summary plot for ACME
}
\description{
This function draw a summary plot of ACME (average causal mediation effect)
}
\examples{

# see univariate_mediation example

}
\author{
Basile Jumentier
}
