% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmed.R
\name{plot_summary_DMR}
\alias{plot_summary_DMR}
\title{Summary plot for univariate_mediation_DMR function}
\usage{
plot_summary_DMR(res_univariate_mediation_DMR, res_DMR_built)
}
\arguments{
\item{res_univariate_mediation_DMR}{result object from res_univariate_mediation_DMR() function}

\item{res_DMR_built}{result object from res_DMR_built() function}
}
\value{
Summary plot
}
\description{
This function draw a summary plot of the mediation analysis
}
\examples{

# see univariate_mediation_DMR example

}
\author{
Basile Jumentier
}
