% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmed.R
\name{plot_summary_med}
\alias{plot_summary_med}
\title{Summary plot for univariate_mediation function}
\usage{
plot_summary_med(res_univariate_mediation)
}
\arguments{
\item{res_univariate_mediation}{result object from univariate_mediation() function}
}
\value{
Summary plot
}
\description{
This function draw a summary plot of the mediation analysis
}
\examples{

# see univariate_mediation example

}
\author{
Basile Jumentier
}
