% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmed.R
\name{DMR_search}
\alias{DMR_search}
\title{DMR_search Find DMR}
\usage{
DMR_search(chr, start, end, pval, cpg, ...)
}
\arguments{
\item{chr}{chromosomes}

\item{start}{chromosomal position of markers (start)}

\item{end}{chromosomal position of markers (end)}

\item{pval}{pValues for each markers, from the max2 function}

\item{cpg}{name of each markers}

\item{...}{argument of the combp function from the ENmix package}
}
\value{
A set of selected DMRs.
}
\description{
To identify differentially methylated regions using a modified comb-p method.
}
\details{
The function will use a modified comb-p method to identify
differentially methylated regions (DMRs).
}
\examples{

}
\author{
Basile Jumentier
}
